allow emitter attribute to be specified for Event object created by emit()
It should be possible for a class to completely encapsulate an Emitter
object and act like one, e.g.

package Node::Emitter {
    use Moo;
    with 'Beam::Emitter';
}

package Node {

    use Moo;
    has _emitter => (
        is        => 'ro',
        init_args => undef,
        default   => sub { Node::Emitter->new },
        handles   => [qw< emit subscribe unsubscribe emit_args >],
    );
}

Unfortunately, when emit() constructs the Event object to be passed
to the callback, it hardwires the Node::Emitter object in the
"emitter" attribute.

This changeset allows the encapsulating class to specify an alternate
value for the emitter attribute:

package Node::Event {

    use Types::Standard qw(:all);
    use Moo;

    has emitter => (
        is       => 'ro',
        isa      => InstanceOf['Node'],
        required => 1,
    );
}

package Node {

    use Moo;
    has _emitter => (
        is        => 'ro',
        init_args => undef,
        default   => sub { Node::Emitter->new },
        handles   => [qw< emit_args on un subscribe unsubscribe >],
    );

    sub emit {

        my ( $self, $event ) = @_;

        $self->_emitter->emit(
            $event,
            class   => 'Node::Event',
            emitter => $self,
            @_
        );
    }
}

so that callbacks will receive a reference to the Node object, rather
than the encapsulated _emitter object.