fix on/un aliases ignoring modified sub/unsub
If we modified `subscribe` or `unsubscribe` to add more behavior, but
the user used `on` or `un` to subscribe or unsubscribe, our modified
methods would be ignored. This violates inheritance practices.

Now we use proper method resolution to forward to `subscribe` or
`unsubscribe`. It's a little slower, but it behaves correctly.

Fixes #18